+++ lib/to_source/emitter/concat_arguments.rb
@@ -17,12 +17,27 @@ module ToSource
       #
       def dispatch
         emit('[')
-        run(Util::DelimitedBody, node.array.body)
+        emit_array_body
         emit(', *')
         visit(rest)
         emit(']')
       end
 
+      # Emit array body
+      #
+      # @return [undefined]
+      #
+      # @api private
+      #
+      def emit_array_body
+        array = node.array
+        case array
+        when Rubinius::AST::ArrayLiteral
+          run(Util::DelimitedBody, array.body)
+        else
+          visit(array)
+        end
+      end
     end
   end
 end